refactor(#6): reorganize pages/components structure, normalize naming, and update all import paths without logic changes
