+++ src/App.tsx
@@ -1,15 +1,15 @@
 import { useCallback, useEffect, useMemo, useState } from 'react'
 import { Navigate, Route, Routes, useLocation, type Location } from 'react-router-dom'
-import CursorTrail from './components/CursorTrail'
-import InitialLoader from './components/InitialLoader'
-import Navbar from './components/Navbar'
-import PageTransition from './components/PageTransition'
-import SmoothScroll from './components/SmoothScroll'
+import Navbar from './components/layout/Navbar/Navbar'
+import SmoothScroll from './components/navigation/SmoothScroll/SmoothScroll'
+import CursorTrail from './components/overlay/CursorTrail/CursorTrail'
+import InitialLoader from './components/overlay/InitialLoader/InitialLoader'
+import PageTransition from './components/overlay/PageTransition/PageTransition'
 import motionBlurHeroOptimized from './assets/images/motion-blur-hero-optimized.jpg'
-import FreetimePage from './pages/freetime'
-import IndexPage from './pages/index'
-import InfoPage from './pages/info'
-import WorkPage from './pages/work'
+import FreetimePage from './pages/FreetimePage'
+import IndexPage from './pages/IndexPage'
+import InfoPage from './pages/InfoPage'
+import WorkPage from './pages/WorkPage'
 import './App.css'
 
 type AppRoutesProps = {
