+++ src/components/navigation/SmoothScroll/SmoothScroll.tsx
