+++ src/components/sections/GallerySection/GallerySection.css
