+++ src/pages/IndexPage.tsx
@@ -1,7 +1,7 @@
 import { useLayoutEffect, useRef } from 'react'
-import GallerySection from '../components/GallerySection'
-import SiteFooter from '../components/SiteFooter'
 import { Link } from 'react-router-dom'
+import SiteFooter from '../components/layout/SiteFooter/SiteFooter'
+import GallerySection from '../components/sections/GallerySection/GallerySection'
 import personalPic from '../assets/images/personalpicSW-optimized.jpg'
 
 function IndexPage() {
